refactor: consolidate exported types by purpose
Consolidate runner input and runner result types in isolated files. Previously, we had a mish-mash of how the types are structured: some were exported from isolated files, some were bundled together. Now, the grouping is based by the runner attribution: input options are located in one file, execution result types in another.
